--- utils/src/files.ts
@@ -1,7 +1,0 @@
-export function createImageFileName(path: string, type: string) {
-  let newPath = path;
-  if (path.startsWith('/')) {
-    newPath = `.${path}`;
-  }
-  return `${newPath}/${type}.png`;
-}
